Added partial bike lane data transformation
This is a preliminary version of the `bikelane_data_transformations`. I have been working on getting the `street_name` column clean & complete. 

As I asked in the morning meeting, I am stuck with what to do with the remaining missing values. I implemented the reverse geocoding method and that worked for the most part, but I still have 264 missing values for the `street_name` column, I did a quick visualisation and found that there are 6 main/tightly clustered missing values with the other missing values scattered between those clusters so maybe I can write a script to assign names to the missing values based on the closest documented street name? 

(You can see all of this information in section 7 of this notebook).

I still need to address the the other missing values, ie:
1. For the missing length values, I am going to use the `geometry`, `longitude`& `latitude`columns to work backs and calculate the lengths. 
2. The percentages of missing values for `cycleway`, `condition`, `lane_type` are all over 85% so I am not sure if I should exclude them from the final table altogether or if I should fill the missing values with something like “undefined” or “none”.